+++ webpack.config.js
@@ -24,8 +24,5 @@ module.exports = {
   output: {
     filename: "script.js",
     path: path.resolve(__dirname, "docs"),
-    environment: {
-      arrowFunction: false,
-    },
   },
 };