+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# unreleased
+
+- Drop support for ES5 browsers
+
 # 0.21.0
 
 - Fix typings for oracle id, where it may not be present at the top level in some circumstances