+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # unreleased
 
 - Drop support for ES5 browsers
+- Drop support for node v12
 - Fix issue where some symbols would return incorrect value when calling `getSymbolUrl`
 
 # 0.21.0